[IMP] estate module: create security rules
This adds user groups and record rules in order to prevent users to see
records from other users. This also hides some functionalities in order
to limit access to users that don't require them.
This is necessary to keep control of the actions and information that
users can use and access.